Add HockeyApp support
To use HockeyApp you need to add a Appsterdam-Keys.plist file in your
Application Support folder.

This file is in the .gitignore file in order to prevent it from the
credentials being added to source control.

Contact @spllr or any of the Appsterdam Milan team if you need access to
the HockeApp account.